requirements.txt
aiofiles
aiosqlite
beautifulsoup4
colorlog
flask
flask-socketio
imageio
isort
pyyaml
pyzipper
requests
ruamel.yaml
toml
